[RES] conflict?

+++ ft_memchr.c
@@ -6,12 +6,14 @@
 /*   By: juggorr <juggorr@gmail.com>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/22 12:12:32 by juggorr           #+#    #+#             */
-/*   Updated: 2023/12/06 22:48:52 by juggorr          ###   ########.fr       */
+/*   Updated: 2023/12/08 12:18:15 by juggorr          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-void	*ft_memchr(const void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	idx;
+	size_t			idx;
 	unsigned char	uc;
 	unsigned char	*tmp;
