+++ ft_memset.c
@@ -6,12 +6,14 @@
 /*   By: junghopa <juhnhopa@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/12 12:07:54 by junghopa          #+#    #+#             */
-/*   Updated: 2023/12/05 19:14:03 by juggorr          ###   ########.fr       */
+/*   Updated: 2023/12/08 12:42:17 by juggorr          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-void	*ft_memset(void *b, int c, unsigned int len)
+#include "libft.h"
+
+void	*ft_memset(void *b, int c, size_t len)
 {
-	unsigned int	idx;
+	size_t			idx;
 	unsigned char	uc;
 	unsigned char	*tmp;
